+++ src/resources/ResourceQuotas/ResourceQuotaDetails.tsx
@@ -1,13 +1,43 @@
 import { ResourceDetails } from 'shared/components/ResourceDetails/ResourceDetails';
 import { ResourceDescription } from '.';
 import ResourceQuotaCreate from './ResourceQuotaCreate';
-import ResourceQuotaLimits, { ResourceQuotaProps } from './ResourceQuotaLimits';
+import ResourceQuotaLimits from './ResourceQuotaLimits';
 import { UI5Panel } from 'shared/components/UI5Panel/UI5Panel';
 import { LayoutPanelRow } from 'shared/components/LayoutPanelRow/LayoutPanelRow';
 import { useTranslation } from 'react-i18next';
 import { Tokens } from 'shared/components/Tokens';
 import { Text, Title } from '@ui5/webcomponents-react';
 import { spacing } from '@ui5/webcomponents-react-base';
+
+export type ResourceQuotaProps = {
+  kind: string;
+  apiVersion: string;
+  metadata: {
+    name: string;
+    namespace: string;
+  };
+  spec: {
+    scopes?: string[];
+    hard: {
+      [key: string]: string;
+    };
+    scopeSelector?: {
+      matchExpressions: {
+        scopeName: string;
+        operator: string;
+        values?: string[];
+      }[];
+    };
+  };
+  status: {
+    hard: {
+      [key: string]: string;
+    };
+    used: {
+      [key: string]: string;
+    };
+  };
+};
 
 export default function ResourceQuotaDetails(props: any) {
   const { t } = useTranslation();
